test(preview): isolate clipboard cases from artifact transport fixtures
